requirements.txt
scalingqa @  git+https://github.com/KNOT-FIT-BUT/scalingQA.git@master#egg=scalingqa
future==0.18.2
h5py==2.10.0
jsonlines==1.2.0
numpy==1.19.2
progressbar==2.5
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
six==1.15.0
torchtext==0.4.0
tqdm==4.50.0
transformers==4.3.3
urllib3==1.25.10
windPyUtils==1.0.6
windPyTorchUtils==1.0.2
